% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{combineFaFiles}
\alias{combineFaFiles}
\title{combineFaFiles combines multiple fa files into one file, allowing adding grams and labels to the sequence title.}
\usage{
combineFaFiles(fafiles, ofile, addGram = FALSE, label = NULL, verbose = FALSE)
}
\arguments{
\item{fafiles}{fa file names, or file names with only sequences (each line one sequence).}

\item{ofile}{output fa file name.}

\item{addGram}{TRUE to add the gram in the sequence file name to the sequence title. Default is FALSE.}

\item{label}{If provided (e.g., ":1"), then add label to the end of the sequence title. Default is NULL.}

\item{verbose}{if TRUE, then show msg about seq numbers in fa files. Default is FALSE.}
}
\value{
number of sequences in `ofile`.
}
\description{
combineFaFiles combines multiple fa files into one file, allowing adding grams and labels to the sequence title.
}
\examples{
\dontrun{
## combine fa files
combineFaFiles(fafiles=list.files('intron_realPA_Arich_seq', full.names=TRUE),
               ofile='../modelDataSplits/intron_realPA_Arich_seq.all.fa',
               addGram = TRUE, label=":1",
               verbose=TRUE)
## combine sequence files (each line one seq)
}
}
