% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{combineFaFiles_fast}
\alias{combineFaFiles_fast}
\title{combineFaFiles_fast combines multiple fa files into one file using readLines.}
\usage{
combineFaFiles_fast(fafiles, ofile, verbose = FALSE)
}
\arguments{
\item{fafiles}{fa file names.}

\item{ofile}{output fa file name.}

\item{verbose}{if TRUE, then show msg about seq numbers in fa files. Default is FALSE.}
}
\value{
number of sequences in `ofile`.
}
\description{
combineFaFiles_fast combines multiple fa files into one file using readLines.
}
