% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{compareFaFiles}
\alias{compareFaFiles}
\title{compareFaFiles counts overlapping of sequence titles between two fa files.}
\usage{
compareFaFiles(f1, f2)
}
\arguments{
\item{f1}{the first fa file.}

\item{f2}{the second fa file.}
}
\value{
number of sequences and overlapped sequences.
}
\description{
compareFaFiles counts overlapping of sequence titles between two fa files.
}
\examples{
\dontrun{
compareFaFiles('1.fa','2.fa')
}
}
