% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{formatPathFileName}
\alias{formatPathFileName}
\title{formatPathFileName converts Windows file name to Linux.}
\usage{
formatPathFileName(aPath, addBar = FALSE, noExt = FALSE, onlyFileName = FALSE)
}
\arguments{
\item{aPath}{a string}

\item{addBar}{whether to add / in the end of the string, default is FALSE}

\item{noExt}{default is FALSE}

\item{onlyFileName}{default is FALSE, T to get only filename}
}
\value{
a new string in Linux format
}
\description{
formatPathFileName converts Windows file name to Linux.
}
\examples{
formatPathFileName('data\\\\a.bam') # "data/a.bam"
formatPathFileName('data/a.bam') # "data/a.bam"
formatPathFileName('data/a.bam', addBar=T) # "data/a.bam/"
formatPathFileName('data/a.bam/', addBar=T) # "data/a.bam/"
formatPathFileName('data/a.bam/', addBar=T, onlyFileName = TRUE) # a.bam
formatPathFileName('data/a.bam/', noExt=T, onlyFileName = TRUE) # a
}
