% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getArichIPseqs}
\alias{getArichIPseqs}
\title{getArichIPseqs gets A-rich internal priming (IP) sequences based on given PA-free regions.}
\usage{
getArichIPseqs(
  regionsNoPARDS,
  bsgenome,
  grams = NULL,
  N = NULL,
  outputSeqPre = "IP_Arich_seq"
)
}
\arguments{
\item{regionsNoPARDS}{a file from from getFreePARangesIn..function, which stores regions without any pAs.}

\item{bsgenome}{if plotFA, bsgenome should be provided.}

\item{grams}{a string vector of grams (pA signal) to search. pA sequences containing each gram are saved to one .fa file.}

\item{N}{the number of sequences needed. If NULL, output all.}

\item{outputSeqPre}{the output .fa file is like <outputSeqPre>.AATAAA.fa. Default is IP_Arich_seq.}
}
\value{
output fa file names, and output sequences of each gram (or NOPAS) to a .fa file.
}
\description{
getArichIPseqs gets A-rich internal priming (IP) sequences based on given PA-free regions.
}
\details{
Given regionsNoPARDS (from getFreePARangesIn.. functions) with no PA at all, this function outputs the IP sequence of A-rich.
It takes about 30m to output about 9w sequences (500 sequences/10s) on a normal notebook.
}
\examples{
\dontrun{
getArichIPseqs(utrsNoPARDS='utrsNoPA.RDS', bsgenome, grams=gramsMM,
                 N=NULL, outputSeqPre="IP_Arich_seq")
getArichIPseqs(utrsNoPARDS='utrsNoPA.RDS', bsgenome, grams=NULL,
                 N=2000, outputSeqPre="IP_Arich_seq")
}
}
