% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getArichPAseqs}
\alias{getArichPAseqs}
\title{getArichPAseqs gets A-rich pA sequences within `granges`.}
\usage{
getArichPAseqs(
  paBedFiles,
  granges,
  bsgenome,
  chrs = NULL,
  grams = NULL,
  gramsPriority = NULL,
  outputSeqPre = "realPA_Arich_seq",
  shift = 0
)
}
\arguments{
\item{paBedFiles}{one or more BED files storing PAs.}

\item{granges}{a GRanges object to subset pAs.}

\item{bsgenome}{if plotFA, bsgenome should be provided.}

\item{chrs}{a string vector storing chrs to subset, default is NULL.}

\item{grams}{a string vector of grams (pA signal) to search within -10 to -50 of pAs.
pA sequences containing each gram are saved to one .fa file.
If grams=NULL, then polyA signal is not scanned, just output all A-rich seqs.}

\item{gramsPriority}{same length as grams, to set the priority of searching, e.g., c(1, 2, 3, 3, 3)}

\item{outputSeqPre}{the output .fa file is like <outputSeqPre>.AATAAA.fa.}

\item{shift}{0 not to shift PA, >0 shift to 3' end, <0 shift to 5'end.}
}
\value{
output fa file names, and output sequences of each gram to a .fa file, like <realPA_Arich_seq.AATAAA/NOPAS.fa>.
}
\description{
getArichPAseqs gets A-rich pA sequences within `granges`.
}
\details{
The function first gets reference pAs from `paBedFiles`, subsetting by `chrs`. And then extract A-rich pAs.
}
\examples{
\dontrun{
getArichPAseqs(paBedFiles='Human_hg38.PolyA_DB.bed',
               granges, bsgenome=bsgenome, chrs=chrs,
               grams=gramsMM, gramsPriority=gramsPriority,
               outputSeqPre="realPA_Arich_seq")
}
}
