% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getArichPAseqsInIntron}
\alias{getArichPAseqsInIntron}
\title{getArichPAseqsInIntron gets A-rich pA sequences in introns.}
\usage{
getArichPAseqsInIntron(
  paBedFiles,
  txdb,
  bsgenome,
  chrs = NULL,
  grams = NULL,
  gramsPriority = NULL,
  outputSeqPre = "intron_realPA_Arich_seq",
  shift = 0
)
}
\arguments{
\item{paBedFiles}{one or more BED files.}

\item{txdb}{a TXDB object or a data.frame with columns (seqnames, strand, start, end) to get intron ranges.}

\item{bsgenome}{if plotFA, bsgenome should be provided.}

\item{chrs}{a string vector storing chrs to subset, default is NULL.
If provided, will try to format the seqnames in txdb to make consistent with chrs.}

\item{grams}{a string vector of grams (pA signal) to search. pA sequences containing each gram are saved to one .fa file.}

\item{gramsPriority}{same length as grams, to set the priority of searching, e.g., c(1, 2, 3, 3, 3)}

\item{outputSeqPre}{the output .fa file is like <outputSeqPre>.AATAAA.fa. Default is intron_realPA_Arich_seq.}

\item{shift}{0 not to shift PA, >0 shift to 3' end, <0 shift to 5'end.}
}
\value{
output fa file names, and output sequences of each gram to a .fa file.
}
\description{
getArichPAseqsInIntron gets A-rich pA sequences in introns.
}
\details{
The function first gets reference pAs from `paBedFiles`, subsetting by `chrs`. And then extract A-rich pAs in introns.
}
