% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getDLdata}
\alias{getDLdata}
\title{getDLdata gets a combined fa file from true and false seq dir.}
\usage{
getDLdata(tSplitDir, fSplitDir, outputPre = "./train.T.F", exactMatch = TRUE)
}
\arguments{
\item{tSplitDir}{dir of fa files. This function does not check the label (T/F), just combine two dirs.}

\item{fSplitDir}{dir of fa files.}

\item{outputPre}{output dir and file name prefix, e.g., "./train.T.F" to output files like train.T.F.1/2/...fa.}

\item{exactMatch}{default is TRUE.
If exactMatch=TRUE, then splits number and index in tSplitDir and fSplitDir should be matched.
If it is FALSE, then will rep the split index of tSplitDir/fSplitDir with less splits to make the number of splits equal.}
}
\value{
NULL. Will create the output dir and output the fa files.
}
\description{
getDLdata gets a combined fa file from true and false seq dir.
}
\details{
If exactMatch=TRUE, then splits number and index in tSplitDir and fSplitDir should be matched.
}
\examples{
\dontrun{
getDLdata(tSplitDir='realPA_Arich_seq_train_1000s_10splits',
          fSplitDir='IP_Arich_seq_train_1000s_10splits',
          outputPre="train.1000T.1000F/train.1000T.1000F",
          exactMatch=TRUE)
}
}
