% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getFreePARangesIn3UTR}
\alias{getFreePARangesIn3UTR}
\title{getFreePARangesIn3UTR gets genomic regions including 3'UTRs and downstream regions but free of pAs.}
\usage{
getFreePARangesIn3UTR(
  paBedFiles,
  txdb,
  extUTRLen = 5000,
  extPA = 200,
  outputRds = "utrsNoPA.RDS",
  chrs = NULL,
  plotFA = TRUE,
  N = 5000,
  bsgenome = NULL,
  outputDir = "./"
)
}
\arguments{
\item{paBedFiles}{one or more BED files.}

\item{txdb}{a TXDB object or a data.frame with columns (seqnames, strand, start, end) to get 3UTR ranges.}

\item{extUTRLen}{length to extend 3UTR downstream for searching A-rich positions.
Larger extUTRlen to get more utrsNoPA for subsequent IP scanning.}

\item{extPA}{extend PA regions.}

\item{outputRds}{the output .RDS file name, default is `utrsNoPA.RDS`.}

\item{chrs}{a string vector storing chrs to subset, default is NULL.
If provided, will try to format the seqnames in txdb to make consistent with chrs.}

\item{plotFA}{TRUE to plot base composition for the 3UTR end position.
`N` PAs will be sampled plot each bed file, and output into BED_files_base_profile.pdf.}

\item{N}{number of PAs sampled to plot each BED file.}

\item{bsgenome}{if plotFA, bsgenome should be provided.}

\item{outputDir}{the output dir for .RDS file and other .freq/.pdf files (if plotFA=TRUE)}
}
\value{
a GRanges object, and output the .RDS file to `outputRds`.
}
\description{
getFreePARangesIn3UTR gets genomic regions including 3'UTRs and downstream regions but free of pAs.
}
\details{
The output GRanges contains 3UTRs and their downstream 5000bp regions, but withou any pAs.
These ranges do not include potential PA or near the end of 3UTR [-200,+200], which are completely PA-free regions.
Each region in the GRanges is at least 500 nt.
}
