% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getFreePARangesInNon3UTRs}
\alias{getFreePARangesInNon3UTRs}
\title{getFreePARangesInNon3UTRs gets PA-free regions of 5UTR/CDS/intron/intergenic.}
\usage{
getFreePARangesInNon3UTRs(
  paBedFiles,
  txdb,
  region = "all",
  chrs = NULL,
  tileWidth = 1000,
  outputRds = "non3UTRsNoPA.RDS"
)
}
\arguments{
\item{paBedFiles}{one or more BED files.}

\item{txdb}{a TXDB object.}

\item{region}{genomic region to output, including all, cds, intron, 5utr, intergenic.}

\item{chrs}{a string vector storing chrs to subset, default is NULL.
If provided, will try to format the seqnames in txdb to make consistent with chrs.}

\item{tileWidth}{tile larger region to smaller pieces.}

\item{outputRds}{the output .RDS file name, default is `non3UTRsNoPA.RDS`.
If region=all, then the output RDS will be like <non3UTRsNoPA_5utr/intron/intergneic/cds.RDS>.}
}
\value{
a GRanges object, and output the .RDS file(s) to `outputRds`.
}
\description{
getFreePARangesInNon3UTRs gets PA-free regions of 5UTR/CDS/intron/intergenic.
}
\details{
The output GRanges contains only ranges without any pAs.
Each region in the GRanges is at least 500 nt.
For long ranges, will tile into `tileWidth` bp ranges to split long regions.
}
