% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getNseqs}
\alias{getNseqs}
\title{getNseqs gets a combined fa file for each split following a given grams distribution.}
\usage{
getNseqs(seqDir, N, nsplits = 10, outputPre, perc = NULL, seed = 123)
}
\arguments{
\item{seqDir}{dir of fa files with each gram a file, e.g., AATAAA.fa.}

\item{N}{the number of sequences for each split (fa file).}

\item{nsplits}{number of splits (fa files), default is 10.}

\item{outputPre}{output dir and file name prefix, e.g., "./deepDataTrain" to output files like deepDataTrain.split.1/2/...fa.}

\item{perc}{a vector of gram proportion or gram counts, which stores the fraction of each gram in `seqDir`.
The names of the `perc` vector should contain all grams in `seqDir`.
If perc=NULL, then will calculate the fraction from `seqDir`.
If perc is count, then will calculate the percentages.}

\item{seed}{a random seed for shuffle the order of sequences of each fa file in `seqDir`.}
}
\value{
a file list based on `outputPre`, e.g., <outputPre>.split.1.fa
}
\description{
getNseqs gets a combined fa file for each split following a given grams distribution.
}
\details{
getNseqs tries to split given fa files without replacement.
However, when the sequence number of a gram is less than the needed number for all splits (defined by `nsplits`), the sequences of that gram in different splits may be duplicated
To meet the given `N` in each split, the fa file of the gram with largest sequence number will be used to supplement.
This function requires that the number of sequences in each fa file is enough for at least one split, otherwise, please reduce `N` or `nsplits`.
}
\examples{
\dontrun{
## split to 10 splits each with 1000s
files=getNseqs(seqDir='realPA_Arich_seq_test/',
               N=1000, nsplits=10,
               outputPre="./realPA_Arich_seq_test_1000s_10splits",
               perc=NULL)
plotATCGforFAfile (faFiles=files, ofreq=FALSE, opdf=T, refPos=101,
             filepre='realPA_Arich_seq_test_1000s_10splits', mergePlots=TRUE)

## split IP files, using perc from real PA files
perc=statCntFas(path='polyAseqTrap/modelData/', filePre='realPA_Arich_seq.')
## split to 2 splits each with 4000s
files=getNseqs(seqDir='IP_Arich_seq_train/',
               N=4000, nsplits=2,
               outputPre="./IP_Arich_seq_train_4000s_2splits",
               perc=perc)

## randomly generate 1 split with random seed,
## which can be run multiple times with different seeds
## to get different splits even there are some overlapping.
files=getNseqs(seqDir='realPA_Arich_seq_test/',
               N=10000, nsplits=1,
               outputPre="realPA_Arich_seq_test_10000s",
               perc=NULL, seed=9999)
}
}
