% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getTxdb5UTRs}
\alias{getTxdb5UTRs}
\title{getTxdb5UTRs gets 5UTR ranges in TXDB.}
\usage{
getTxdb5UTRs(txdb, chrs = NULL)
}
\arguments{
\item{txdb}{a TXDB object , or a data.frame with at least columns: seqnames, start, end ,strand, to get 5UTR ranges.}

\item{chrs}{a string vector storing chrs to subset, default is NULL.
If provided, will try to format the seqnames in txdb to make consistent with chrs.}
}
\value{
a GRanges of 5UTRs.
}
\description{
getTxdb5UTRs gets 5UTR ranges in TXDB.
}
