% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getTxdbIntergenics}
\alias{getTxdbIntergenics}
\title{getTxdbIntergenics gets intergenic ranges in TXDB.}
\usage{
getTxdbIntergenics(txdb, chrs = NULL, extLen = 5000)
}
\arguments{
\item{txdb}{a TXDB object , or a data.frame with at least columns: seqnames, start, end ,strand, to get intergenic ranges.}

\item{chrs}{a string vector storing chrs to subset, default is NULL.
If provided, will try to format the seqnames in txdb to make consistent with chrs.}

\item{extLen}{length to extend transcript downstream to narrow intergenic ranges.}
}
\value{
a GRanges of intergenics, whose both strands not within any transcripts+extLen.
}
\description{
getTxdbIntergenics gets intergenic ranges in TXDB.
}
\examples{
\dontrun{
itgs=getTxdbIntergenics(txdb)
}
}
