% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{getTxdbUTRs}
\alias{getTxdbUTRs}
\title{getTxdbUTRs gets 3UTR ranges in TXDB.}
\usage{
getTxdbUTRs(
  txdb,
  extLen = 5000,
  plotFA = TRUE,
  N = 5000,
  bsgenome = NULL,
  chrs = NULL,
  outputDir = NULL
)
}
\arguments{
\item{txdb}{a TXDB object, or a data.frame with at least columns: seqnames, start, end ,strand}

\item{extLen}{length to extend 3UTR downstream for searching A-rich positions}

\item{plotFA}{TRUE to plot base composition for the 3UTR end position.
`N` PAs will be sampled plot each bed file, and output into BED_files_base_profile.pdf.}

\item{N}{number of PAs sampled to plot each BED file.}

\item{bsgenome}{if plotFA, bsgenome should be provided.}

\item{chrs}{a string vector storing chrs to subset, default is NULL.
If provided, will try to format the seqnames in txdb to make consistent with chrs.}

\item{outputDir}{output dir for .freq, .fa and .pdf files (if plotFA=TRUE). NULL to output to the path of bedfiles.}
}
\value{
a list of utrsPA and utrsExt.
}
\description{
getTxdbUTRs gets 3UTR ranges in TXDB.
}
\details{
This function will output a list of utrsPA and utrsExt, which are used as reference regions for geting IP and real pAs.
}
