% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{plotFaDeepRes}
\alias{plotFaDeepRes}
\title{plotFaDeepRes plots single nucleotide profile and output fasta files for TP/FP/TN/FN.}
\usage{
plotFaDeepRes(deepResCsv, fafile, outputFa = NULL)
}
\arguments{
\item{fafile}{the fa file corresponding to deepResCsvs}

\item{outputFa}{NULL for not outputing fa but just plotting. Otherwise, the prefix for the output fa files.}

\item{deepResCsvs}{the csv file name from DeepIP_test.py.}
}
\value{
fa file names.
Output fasta file as follows:<outputFa or deepResCsv>.FN10.fa means that 1 is judged as 0, indicating FN.
}
\description{
plotFaDeepRes plots single nucleotide profile and output fasta files for TP/FP/TN/FN.
}
\examples{
\dontrun{
plotFaDeepRes('human_on_chlamy.csv',
              'chlamy.fa',
              outputFa='fa/human_on_chlamy')
}
}
