% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{readBedFiles2GR}
\alias{readBedFiles2GR}
\title{readBedFiles2GR reads BED files of pA sites and removes duplicates.}
\usage{
readBedFiles2GR(
  bedfiles,
  chrs = NULL,
  plotFA = TRUE,
  N = 5000,
  bsgenome = NULL,
  outputDir = NULL
)
}
\arguments{
\item{bedfiles}{one or more PA/BED files without header, with >=4 columns or =3 columns (chr/strand/start).}

\item{chrs}{a string vector storing chrs to subset, default is NULL.}

\item{plotFA}{TRUE to plot base composition for each BED file.
`N` PAs will be sampled plot each bed file, and output into BED_files_base_profile.pdf.}

\item{N}{number of PAs sampled to plot each BED file.}

\item{bsgenome}{if plotFA, bsgenome should be provided.}

\item{outputDir}{output dir for .freq, .fa and .pdf files (if plotFA=TRUE). NULL to output to the path of bedfiles.}
}
\value{
GRanges storing pAs.
}
\description{
readBedFiles2GR reads BED files of pA sites and removes duplicates.
}
\details{
This function will check the chr name consistency, and plot base composition for each BED file if plotFA=TRUE.
}
