% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{resetNonPATitle}
\alias{resetNonPATitle}
\title{resetNonPATitle resets sequence title in the fa file to make sure the title reflects the position of the sequence}
\usage{
resetNonPATitle(faFiles, sufix = "_ts")
}
\arguments{
\item{faFiles}{fa file names.
Such fa files are sequences like "nonPA-pos", or "IP=".}

\item{sufix}{output file name be like <faFiles><_ts>.fa}
}
\value{
output file names.
}
\description{
resetNonPATitle resets sequence title in the fa file to make sure the title reflects the position of the sequence
}
