% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{split2TrainTest}
\alias{split2TrainTest}
\title{split2TrainTest randomly splits fa files (each gram one file) to two folders (e.g., train and test)}
\usage{
split2TrainTest(
  path,
  filePre,
  dir1 = "train",
  dir2 = "test",
  per1 = 0.5,
  label = ":1"
)
}
\arguments{
\item{path}{path of the fa files}

\item{filePre}{prefix of the fa file names, e.g., for file "realPA_Arich_seq.NOPAS.fa", the filepre is "realPA_Arich_seq.".
The gram should be present in the file name, that is the file name is <filePre>.gram.fa.}

\item{dir1}{like 'dir1' or '../xx/dir2'. split `per1` percentage of fa files to dir1. If dir is not exists, will be created.}

\item{dir2}{split remaining (`1-per1`) percentage of fa files to dir2}

\item{per1}{percentage of sequences to be extracted into dir1, default is 0.5}

\item{label}{add label to the sequence title, default is ":1".}
}
\value{
NULL, output files to dir1 and dir2.
The gram (file name) and the `label` will be added to the sequence title, like <title>;AATAAA:1.
}
\description{
split2TrainTest randomly splits fa files (each gram one file) to two folders (e.g., train and test)
}
\examples{
\dontrun{
split2TrainTest(path='polyAseqTrap/modelData/',
                filePre='realPA_Arich_seq.',
                dir1='realPA_Arich_seq_train', dir2='realPA_Arich_seq_test',
                per1=0.7, label=":1")
}
}
