% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{split2TrainTestByFiles}
\alias{split2TrainTestByFiles}
\title{split2TrainTestByFiles splits each of fafiles to dir1 (per1) and dir2 (1-per1), adding label to the title.}
\usage{
split2TrainTestByFiles(
  fafiles,
  dir1 = "train",
  dir2 = "test",
  per1 = 0.5,
  label = ":1"
)
}
\arguments{
\item{fafiles}{fa file names.}

\item{dir1}{like 'dir1' or '../xx/dir2'. split `per1` percentage of fa files to dir1. If dir is not exists, will be created.}

\item{dir2}{split remaining (`1-per1`) percentage of fa files to dir2}

\item{per1}{percentage of sequences to be extracted into dir1, default is 0.5}

\item{label}{add label to the sequence title, default is ":1".}
}
\value{
NULL, output files to dir1 and dir2.
The `label` will be added to the sequence title, like <title>:1.
}
\description{
split2TrainTestByFiles splits each of fafiles to dir1 (per1) and dir2 (1-per1), adding label to the title.
}
\examples{
\dontrun{
split2TrainTestByFiles(fafiles,
                      dir1='3utr_shiftNonDRSPA_seq_train_per70',
                      dir2='3utr_shiftNonDRSPA_seq_test_per30',
                      per1=0.7, label=":0")
}
}
