% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{statCntFas}
\alias{statCntFas}
\title{statCntFas stats seq number in filename like 'realPA_Arich_seq.ATTATA.fa', 'NOPAS.fa'}
\usage{
statCntFas(path, filePre = "", readNOPAS = TRUE, justTotal = FALSE)
}
\arguments{
\item{path}{path of the fa files}

\item{filePre}{prefix of the fa file names, e.g., for file "realPA_Arich_seq.NOPAS.fa", the filepre is "realPA_Arich_seq.".
The gram should be present in the file name, that is the file name is <filePre>.gram.fa.}

\item{readNOPAS}{TRUE to read NOPAS file}

\item{justTotal}{FALSE to not count total number}
}
\value{
NULL, output to screen
}
\description{
statCntFas stats seq number in filename like 'realPA_Arich_seq.ATTATA.fa', 'NOPAS.fa'
}
