% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{statDeepRes}
\alias{statDeepRes}
\title{statDeepRes gets metrics for .csv prediction results from DeepIP_test.py.}
\usage{
statDeepRes(deepResCsvs, ofile = NULL, verbose = FALSE, shortFileName = TRUE)
}
\arguments{
\item{deepResCsvs}{a string vector storing one or multiple csv files from DeepIP_test.py.
If the name of the vector is provided, then will use the name as the output row title denoting each csv file.}

\item{ofile}{output .csv or other file to store the metrics table. If it is NULL, then not output.}

\item{verbose}{TRUE to print the metrics table.}

\item{shortFileName}{TRUE to remove common strings in file names of deepResCsvs.}
}
\value{
A dataframe of the metrics, with each row one file of `deepResCsvs` and each column a metric.
Metrics are: TP_11	FP_01	TN_00	FN_10	Sensitivity	Specificity	Precision	Recall	F1	ROC	AUC.
}
\description{
statDeepRes gets metrics for .csv prediction results from DeepIP_test.py.
}
\examples{
\dontrun{
statDeepRes('train.10000T.10000F.1.epoch100_on_test.all.csv',
             ofile='train.10000T.10000F.1.epoch100_on_test.all.stat.txt')
}
}
