% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainDeepIP.R
\name{testDeepIP}
\alias{testDeepIP}
\title{testDeepIP test given input sequences using given trained model}
\usage{
testDeepIP(condaEnv, inTestSeq, inTrainedModel, outTestCsv, seqLabel = "")
}
\arguments{
\item{condaEnv}{the python conda environment}

\item{inTestSeq}{fa file for test, could have labels like (/1, _1, or :1) or (/0, _0, or :0) in the sequence title or not.}

\item{inTrainedModel}{trained hdf5 file}

\item{outTestCsv}{output csv file with five columns: title,score,true_label,predict_label,res=TP/FP/FN/TN}

\item{seqLabel}{define the label of the sequences in `inTestSeq`, could be '' or '0/1','0:1','01','1/0','1:0','10','00','0:0','0/0','11','1/1','1:1','1','0'.}
}
\value{
outTestCsv file name.
}
\description{
testDeepIP test given input sequences using given trained model
}
\examples{
\dontrun{
testDeepIP("C:/Users/LENOVO/anaconda3/envs/DeepPASTA/",
            'test.all.fa',
            'train.10000T.10000F/train.10000T.10000F.1.epoch100.hdf5',
            'train.10000T.10000F.1.epoch100_ON_test.all.csv'
            '')
# metrics
statDeepRes('train.10000T.10000F.1.epoch100_ON_test.all.csv',
            ofile='train.10000T.10000F.1.epoch100_ON_test.all.stat.csv')

# TP/FP fa profile
plotFaDeepRes('train.10000T.10000F.1.epoch100_ON_test.all.csv',
              fafile='test.all.fa')
# grams metrics
stats=statDeepResByGrams('train.10000T.10000F.1.epoch100_ON_test.all.csv',
         ofile='train.10000T.10000F.1.epoch100_ON_test.all.statByGrams.csv',
                         plot=TRUE)
}
}
