% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainDeepIP.R
\name{trainDeepIP}
\alias{trainDeepIP}
\title{trainDeepIP trains DeepIP model for given input sequences}
\usage{
trainDeepIP(condaEnv, inTrainSeq, outTrainedModel, epoch = 100, seqLabel = "")
}
\arguments{
\item{condaEnv}{the python conda environment}

\item{inTrainSeq}{fa file containing positive (/1, _1, or :1) and negative (end with /0, _0, or :0) sequences}

\item{outTrainedModel}{output hdf5 file}

\item{epoch}{iteration number, default is 100,}

\item{seqLabel}{define the label of the sequences in `inTrainSeq`, could be '' or 0/1, 1/0, 0:1, 1:0, 01, 10}
}
\value{
outTrainedModel file name.
}
\description{
trainDeepIP trains DeepIP model for given input sequences
}
\examples{
\dontrun{
trainDeepIP("C:/Users/LENOVO/anaconda3/envs/DeepPASTA/",
            'train.10000T.10000F/train.10000T.10000F.1.fa',
            'train.10000T.10000F/train.10000T.10000F.1.epoch100.hdf5',
            100)
}
}
