% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepIP_funclib.R
\name{trimPAseqs}
\alias{trimPAseqs}
\title{trimPAseqs trims each sequence in fa files with start~stop}
\usage{
trimPAseqs(
  faFiles,
  start = NULL,
  stop = NULL,
  sufix = "_trim",
  ofiles = NULL,
  outputDir = NULL,
  basePlot = FALSE
)
}
\arguments{
\item{faFiles}{fa file names.}

\item{start}{start position for all sequences.}

\item{stop}{stop position for all sequences, final sequence will be from start(include)..end (include).}

\item{sufix}{output file name be like <faFiles><_trim>.fa}

\item{ofiles}{must the same length as faFiles. If provided, then will ignore sufix.}

\item{outputDir}{if NULL, then output each file in the dir of corresponding `faFiles`.}

\item{basePlot}{if TURE, to plot first 6 output files' base compositions.}
}
\value{
output file names.
}
\description{
trimPAseqs trims each sequence in fa files with start~stop
}
\examples{
\dontrun{
## start=21, stop=220 to trim 200nt from a 241nt seq with PA in 101
trimPAseqs('3utr_nonPA_seq_train_per70_deepPASTA_5000T5000F_train.fa',
           sufix='_trim200PA101',
           start=21, stop=220, basePlot=TRUE)
}
}
